Expondo atributos para no jsp e corrigindo problema de cast (integer x long)

+++ casadocodigo/src/main/java/br/com/casadocodigo/loja/controllers/ProdutosController.java
@@ -76,7 +76,7 @@ public class ProdutosController {
 	}
 	
 	@RequestMapping("/detalhe/{id}")
-	public ModelAndView detalhe(@PathVariable("id") long id) {
+	public ModelAndView detalhe(@PathVariable("id") Integer id) {
 		ModelAndView modelAndView = new ModelAndView("produtos/detalhe");
 		Produto produto = produtoDao.find(id);
